pyproject.toml
# Created: 2024-07
# Created by: Basile El Azhari
# Maintained by: Basile El Azhari
# Contact: basile.elazhari@ekimetrics.com, https://www.linkedin.com/in/basile-el-azhari/

[tool.poetry]
name = "ezazml"
version = "0.1.0"
description = "An easy way to deploy your machine learning models on Azure Machine Learning"
authors = ["basile el azhari <hitchhikesaver@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
pandas = "^2.2.2"
azure-ai-ml = "1.18.0"
azure-identity = "1.17.1"
azureml-fsspec = "^1.3.1"
mltable = "^1.6.1"
azureml-dataprep = { version = "^5.1.0", extras = ["pandas"] }
click = "^8.1.7"
GitPython = "^3.1.43"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.2"
pre-commit = "^3.6.2"
jupyter = "^1.0.0"
dotenv-cli = "^3.3.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.0.2"
pytest-cov = "^3.0.0"
pytest-mock = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ezazml = "ezazml.cli.cli:cli"

[tool.bandit]
exclude_dirs = ["**/test_*.py", "*/test_*.py"]
tests = ["B201"]
skips = ["B101", "B601"]
